split page files
