+++ src/pages/Pool/styled.tsx
@@ -1,7 +1,13 @@
 import { Text } from 'rebass';
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
+import { Box as RebassBox } from 'rebass'
+import Switch from 'react-switch'
 
 import { LoadingRows as BaseLoadingRows } from 'components/Loader/styled';
+import { AutoColumn } from 'components/Column';
+import { AutoRow, RowBetween } from 'components/Row';
+import { AlertTriangle } from 'react-feather';
+import { ButtonPrimary } from 'components/Button';
 
 export const Wrapper = styled.div`
   position: relative;
@@ -73,3 +79,223 @@ export const LoadingRows = styled(BaseLoadingRows)`
     margin-bottom: 2em;
   }
 `;
+
+export const PageWrapper = styled(AutoColumn)`
+  padding: 0px 8px 0px;
+  max-width: 1020px;
+  width: 100%;
+
+  @media (max-width: ${({ theme }) => `${theme.breakpoint.md}px`}) {
+    padding-top: 20px;
+  }
+`
+
+export const TitleRow = styled(RowBetween)`
+  color: ${({ theme }) => theme.neutral2};
+  @media (max-width: ${({ theme }) => `${theme.breakpoint.sm}px`}) {
+    flex-wrap: wrap;
+    gap: 12px;
+    width: 100%;
+    padding-left: 12px;
+  }
+`
+
+export const ButtonRow = styled(AutoRow)``
+
+export const ErrorContainer = styled.div`
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  margin: auto;
+  min-height: 25vh;
+  @media (max-width: ${({ theme }) => `${theme.breakpoint.md}px`}) {
+    padding: 0px 52px;
+  }
+
+  @media (max-width: ${({ theme }) => `${theme.breakpoint.sm}px`}) {
+    padding: 0px 52px;
+  }
+`
+
+const IconStyle = css`
+  width: 48px;
+  height: 48px;
+  margin-bottom: 0.5rem;
+`
+
+export const NetworkIcon = styled(AlertTriangle)`
+  ${IconStyle}
+`
+
+export const ResponsiveButtonPrimary = styled(ButtonPrimary)`
+  font-family: 'DM Sans';
+  border-radius: 8px;
+  font-size: 16px;
+  padding: 6px 8px;
+  width: 175px;
+  margin-left: auto;
+  height: 38px;
+  @media (max-width: ${({ theme }) => `${theme.breakpoint.sm}px`}) {
+    width: 132px;
+  }
+`
+
+export const ResponsiveButtonTabs = styled(ButtonPrimary) <{ secondary: boolean; active: boolean }>`
+  font-family: 'DM Sans';
+  border-radius: 4px;
+  font-size: 16px;
+  padding: 6px 8px;
+  background: ${({ theme, active }) => (!active ? 'transparent' : theme.jediWhite)};
+  box-shadow: 0px 3.079px 13.856px 0px rgba(154, 146, 210, 0.3) inset,
+    0px 0.77px 30.791px 0px rgba(227, 222, 255, 0.2) inset;
+  color: ${({ theme, active }) => (!active ? theme.jediWhite : theme.jediPink)};
+  width: 121px;
+  margin-left: 0;
+  height: 38px;
+  &:hover {
+    background: ${({ theme, active }) => (!active ? 'transparent' : theme.jediWhite)};
+    color: ${({ theme, active }) => (!active ? theme.jediWhite : theme.jediPink)};
+  }
+  &:active {
+    background: ${({ theme, active }) => (!active ? 'transparent' : theme.jediWhite)};
+    color: ${({ theme, active }) => (!active ? theme.jediWhite : theme.jediPink)};
+  }
+  &:focus {
+    background: ${({ theme, active }) => (!active ? 'transparent' : theme.jediWhite)};
+    color: ${({ theme, active }) => (!active ? theme.jediWhite : theme.jediPink)};
+  }
+  @media (max-width: ${({ theme }) => `${theme.breakpoint.sm}px`}) {
+    width: 50%;
+    margin-bottom: 10px;
+  }
+`
+export const MainContentWrapper = styled.main<{ isWalletConnected?: boolean; filteredPositions?: any }>`
+  background-color: ${({ theme, isWalletConnected, filteredPositions }) =>
+    isWalletConnected && filteredPositions ? 'rgba(196, 196, 196, 0.01)' : 'transparent'};
+  border-radius: 8px;
+  box-shadow: ${({ isWalletConnected, filteredPositions }) =>
+    isWalletConnected && filteredPositions
+      ? `0px 0.76977px 30.79088px 0px rgba(227, 222, 255, 0.2) inset,
+        0px 3.07909px 13.8559px 0px rgba(154, 146, 210, 0.3) inset,
+        0px 75.43767px 76.9772px -36.94907px rgba(202, 172, 255, 0.3) inset,
+        0px -63.12132px 52.3445px -49.26542px rgba(96, 68, 144, 0.3) inset`
+      : ''};
+  @media (max-width: ${({ theme }) => `${theme.breakpoint.md}px`}) {
+  }
+`
+
+export const PanelWrapper = styled.div`
+  display: flex;
+  gap: 12px;
+  width: 100%;
+  align-items: start;
+  @media screen and (max-width: 1024px) {
+    flex-direction: column;
+  }
+`
+const panelPseudo = css`
+  :after {
+    content: '';
+    position: absolute;
+    left: 0;
+    right: 0;
+    height: 10px;
+  }
+
+  @media only screen and (min-width: 40em) {
+    :after {
+      content: unset;
+    }
+  }
+`
+
+export const Panel = styled(RebassBox) <{
+  hover?: boolean
+  background?: boolean
+  area?: boolean
+  grouped?: boolean
+  rounded?: boolean
+  last?: boolean
+}>`
+  position: relative;
+  // background-color: ${({ theme }) => theme.advancedBG};
+  border-radius: 8px;
+  padding: 1.25rem;
+  width: 100%;
+  height: 100%;
+  overflow: hidden;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  background: rgba(196, 196, 196, 0.01);
+  box-shadow: 0px 0.77px 30.791px 0px rgba(227, 222, 255, 0.2) inset,
+    0px 3.079px 13.856px 0px rgba(154, 146, 210, 0.3) inset,
+    0px 75.438px 76.977px -36.949px rgba(202, 172, 255, 0.3) inset,
+    0px -63.121px 52.345px -49.265px rgba(96, 68, 144, 0.3) inset;
+
+  :hover {
+    cursor: ${({ hover }) => hover && 'pointer'};
+    border: ${({ hover, theme }) => hover && '1px solid' + theme.bg5};
+  }
+
+  ${(props) => props.background && `background-color: ${props.theme.advancedBG};`}
+
+  ${(props) => (props.area ? `grid-area: ${props.area};` : null)}
+
+  ${(props) =>
+    props.grouped &&
+    css`
+      @media only screen and (min-width: 40em) {
+        &:first-of-type {
+          border-radius: 20px 20px 0 0;
+        }
+        &:last-of-type {
+          border-radius: 0 0 20px 20px;
+        }
+      }
+    `}
+
+  ${(props) =>
+    props.rounded &&
+    css`
+      border-radius: 8px;
+      @media only screen and (min-width: 40em) {
+        border-radius: 10px;
+      }
+    `};
+
+  ${(props) => !props.last && panelPseudo}
+`
+export const PanelTopLight = styled(Panel)`
+  box-shadow: 0px 0.77px 30.791px 0px rgba(227, 222, 255, 0.2) inset,
+    0px 3.079px 13.856px 0px rgba(154, 146, 210, 0.3) inset,
+    0px 75.438px 76.977px -36.949px rgba(202, 172, 255, 0.3) inset,
+    0px -63.121px 52.345px -49.265px rgba(96, 68, 144, 0.3) inset, 0px 5.388px 8.467px -3.079px #fff inset,
+    0px 30.021px 43.107px -27.712px rgba(255, 255, 255, 0.5) inset;
+`
+
+export const PageHeader = styled.div`
+  color: ${({ theme }) => theme.jediWhite};
+  font-family: "Avenir LT Std";
+  font-size: 24px;
+  font-weight: 750;
+  margin-bottom: 20px;
+`
+
+export const OnlyRewardedSwitcherContainer = styled.div`
+display: flex;
+align-items: center;
+gap: 16px;
+margin: 15px 0;
+`
+
+export const OnlyRewardedSwitcherLabel = styled.div`
+font-family: Avenir LT Std;
+font-size: 18px;
+font-weight: 750;
+line-height: 18px;
+text-align: left;
+`
+
+export const OnlyRewardedSwitcher = styled(Switch)``
